Added functionality for Members, create, edit delete, and db migration

+++ Migrations/LibraryContextModelSnapshot.cs
@@ -87,8 +87,8 @@ namespace Bookish.Migrations
                     b.Property<string>("Name")
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("PhoneNumber")
-                        .HasColumnType("int");
+                    b.Property<long>("PhoneNumber")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
