+++ Models/MemberViewModel.cs
@@ -7,6 +7,7 @@ namespace Bookish.Models
         public MemberViewModel()
         {
         }
+
         public MemberViewModel(MemberDbModel memberDb)
         {
             Id = memberDb.Id;
@@ -18,6 +19,6 @@ namespace Bookish.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public int PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
     }
 }
